backend/requirements.txt
# ============================================================================
# Web Framework & Server
# ============================================================================
fastapi==0.122.0
uvicorn==0.38.0
python-multipart==0.0.20
websockets==15.0.1
watchfiles==1.1.1

# ============================================================================
# Database & ORM
# ============================================================================
sqlalchemy==2.0.45
sqlmodel==0.0.27
alembic==1.17.2
sqlalchemy-pytds
psycopg2-binary==2.9.11
psycopg-binary==3.3.2
asyncpg==0.30.0

# ============================================================================
# Vector Store & Embeddings
# ============================================================================
qdrant-client==1.16.2

# ============================================================================
# LLM & AI
# ============================================================================
cohere==5.20.1
google-genai==1.56.0
google-generativeai
google-cloud-bigquery
vanna==0.7.3
pyodbc

# LlamaIndex Core
llama-index==0.14.10
llama-index-vector-stores-qdrant==0.9.0
llama-index-storage-chat-store-postgres==0.3.2

# LlamaIndex LLMs
llama-index-llms-google-genai==0.8.3
llama-index-llms-mistralai
llama-index-llms-ollama
llama-index-llms-openai-like

# LlamaIndex Embeddings
llama-index-embeddings-gemini==0.4.1
llama-index-embeddings-openai==0.5.1
llama-index-embeddings-huggingface==0.6.1

# LlamaIndex Readers
llama-index-readers-file==0.5.5
llama-index-readers-web==0.5.6
llama-index-readers-confluence==0.6.1
llama-index-readers-microsoft-sharepoint==0.6.1

# ============================================================================
# Document Processing
# ============================================================================
pypdf==6.5.0
python-docx==1.2.0
docx2txt==0.8
openpyxl==3.1.5
pandas==2.2.3

# ============================================================================
# Authentication & Security
# ============================================================================
python-jose[cryptography]==3.5.0
email-validator==2.3.0
passlib[bcrypt]==1.7.4

# ============================================================================
# Caching & Background Tasks
# ============================================================================
redis==7.1.0
APScheduler==3.11.1
croniter==6.0.0

# ============================================================================
# Utilities
# ============================================================================
pydantic-settings==2.12.0
aiofiles==25.1.0
tenacity==8.2.3
psutil==7.1.3

# ============================================================================
# Development Tools
# ============================================================================
flake8==7.3.0
black==25.12.0
isort==7.0.0
mypy==1.19.1

# ============================================================================
# Testing
# ============================================================================
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
pytest-cov==7.0.0
httpx==0.28.1

# ============================================================================
# Observability
# ============================================================================
arize-phoenix
llama-index-callbacks-arize-phoenix
openinference-instrumentation-llama-index
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp